move wcs functions from fits_image into wcs_helpers
