+++ scripts/fix_beam.py
@@ -16,7 +16,7 @@ __institute__ = 'Shanghai Astronomical Observatory'
 import sys
 import argparse
 from AegeanTools.fits_interp import load_file_or_hdu
-from AegeanTools.fits_image import fix_aips_header
+from AegeanTools.wcs_helpers import fix_aips_header
 
 
 def search_beam(hdulist):
